desktop: add <release> tag for 2.99.12.
It's better to start early as these need to be translated too and are
now even visible in GIMP itself.

Note that I wrote 2.99.11 in the version attribute, but only to be able
to test this easily during development. It will have to be edited to the
right version at release.