plug-ins: don't set the BigTIFF checkbox insensitive on second try.
Actually I realized that using BigTIFF is not the only valid alternative
here. We can also try to use a better compression algorithm (or any if
"None" were selected) and try again to export as ClassicTIFF.

So don't check the option ourselves and don't set it insensitive. Let
one choose fully what they prefer to do.