cursors: make the "single dot" cursor actually slightly bigger and B&W.
Committer (Jehan)'s note:

We know the goal is for it not to be too visible, but here it was really
hardly visible at all. Also now we can actually surround the hot spot
position.

As for the color, if the cursor is black only, it becomes extra-hard to
see on dark background images (or even totally invisible on full black
background). All cursors are drawn with B&W color on purpose.

Setting Aryeom as author, because I was trying colors on one side (then
it looked like micro-lines), or bigger checkerboard (looked weird) or
some other design where it looked like a circle in one case and a cross
in another. Aryeom proposed this much better design instead, which stays
symmetrical, small and not-weird/too widely different whatever the
background color.