Issue #8005: GIMP depends on gvfs, but it is not listed in INSTALL.in.
It's not a direct dependency and doesn't prevent GIMP from any of its
core logics, but HTTP is a used enough format nowadays that we want to
consider it as mandatory nowadays.

From what I know, `gvfs` might be for Windows only though. This would
explain issues we have with remote URL on other platforms.